Improve AJoin slicing (#97)
* Fix bug that resulted in an exception when a watermark didn't close a window
 * Use proper slicing in StreamASource

+++ engine/src/main/java/de/hpi/des/hdes/engine/shared/join/StreamASource.java
@@ -26,7 +26,7 @@ public class StreamASource<IN, KEY> extends AbstractTopologyElement<Bucket<KEY, 
   private final KeySelector<IN, KEY> keySelector;
   private final Map<Window, Map<KEY, Set<IN>>> state;
 
-  public StreamASource(final int triggerInterval, final KeySelector<IN, KEY> keySelector) {
+  public StreamASource(final long triggerInterval, final KeySelector<IN, KEY> keySelector) {
     this.sliceAssigner = TumblingWindow.ofEventTime(triggerInterval);
     this.keySelector = keySelector;
     this.state = new HashMap<>();
